añade producto a cesta sumando cantidad o añadiendolo

+++ src/database/cestasModelo.js
@@ -46,28 +46,46 @@ const insertProductCesta = (username,idProducto) => {
         return objeto.id === idProducto;
     });
 
-    // const lengthCestas = datos.cestas.length
-
     for (let i=0; i<datos.cestas.length; i++){
         if (datos.cestas[i].username===username){
-            let objetoIdProducto = {
-                id:product.id
-            }
-            datos.cestas[i].productos.push(objetoIdProducto)
-            fs.writeFile(
-                "./src/database/cestas.json",
-                JSON.stringify(datos, null, 2),
-                "utf8",
-                (err) => {
-                    throw new Error("ERROR AL ESCRIBIR");
+
+            let existe = false
+            for (let x=0;x<datos.cestas[i].productos.length;x++){
+                console.log(datos.cestas[i].productos[x].id)
+                console.log(product.id)
+                if (datos.cestas[i].productos[x].id===product.id){
+                    console.log(datos.cestas[i].productos[x].id.cantidad)
+                    existe = true
+                    datos.cestas[i].productos[x].cantidad+=1
+                    fs.writeFile(
+                        "./src/database/cestas.json",
+                        JSON.stringify(datos, null, 2),
+                        "utf8",
+                        (err) => {
+                            throw new Error("ERROR AL ESCRIBIR");
+                        }
+                    );
+                    return datos.cestas[i]
                 }
-            );
-            return datos.cestas[i]
+            }
+            if (!existe){
+                let objetoProducto = {
+                    id:product.id,
+                    cantidad:1
+                }
+                datos.cestas[i].productos.push(objetoProducto)
+                fs.writeFile(
+                    "./src/database/cestas.json",
+                    JSON.stringify(datos, null, 2),
+                    "utf8",
+                    (err) => {
+                        throw new Error("ERROR AL ESCRIBIR");
+                    }
+                );
+                return datos.cestas[i]
+            }
         }
     }
-
-
-
 }
